e761aab2 COMP: various changes for C++11 compatibility

+++ Filtering/Testing/Cxx/TestPolyDataRemoveCell.cxx
@@ -36,7 +36,7 @@ int TestPolyDataRemoveCell(int , char *[])
   points->SetNumberOfPoints(numPoints);
   for(i=0;i<numPoints;i++)
     {
-    double loc[3] = {i, i*i, 0};
+    double loc[3] = {static_cast<double>(i), static_cast<double>(i*i), 0.0};
     points->InsertPoint(i, loc);
     }
   vtkSmartPointer<vtkPolyData> poly = vtkSmartPointer<vtkPolyData>::New();
